re-modularisation pour reutilisations modules hors vue. Ajount fonction survol etats dans viewmode/createCricles.js
